99_アーカイブ/devtools/transcription/requirements.txt
openai-whisper


